Fix some "int" to "auto"

+++ Encoder.cpp
@@ -5,7 +5,7 @@
 #include "Encoder.h"
 
 
-const unsigned int Encoder::MEDIAN = 30000;
+const uint16_t Encoder::MEDIAN = 30000;
 const float Encoder::PULSE_L = 0.021475781;
 const float Encoder::PULSE_R = 0.021475781;
